Issue #28595 refactor methods to QueryInfo
